Fixes LIBCLOUD-931 - cloudfiles download_object_as_stream
When downloading an object as a stream `download_object_as_stream`
in storage/drivers/cloudfiles.py a generator object should be
returned. Currently a rerquest/response object is returned
in resulting in a HttpLibResponseProxy error.

This commit alters the callback_kwargs to use a
requests/response.iterator resulting in an actual iterator
being returned.